requirements.txt
Django==5.0
django-admin-sortable2==2.1.10
python-dotenv==1.0.1
djangorestframework==3.14.0
pillow==10.2.0
django-colorfield==0.11.0
django-ckeditor==6.7.1
python-jose==3.3.0
pydantic-settings==2.2.0
gunicorn==22.0.0
celery==5.3.6
django-hosts==6.0
django-cors-headers==4.3.1
cryptography==41.0.7
libsass==0.23.0
django_compressor==4.5
django-sass-processor==1.4.1
channels==4.1.0
channels-redis==4.2.0
django-qr-code==4.1.0
daphne==3.0.2
unidecode==1.3.8
russian-names==0.1.2
asgiref==3.8.0
